third_party/WebKit/LayoutTests/fast/events/scroll-in-scaled-page-with-overflow-hidden-expected.txt

This tests that a usually overflow: hidden viewport should be scrollable when scaled. Otherwise, you can't get to content you would have been able to get to had you not been zoomed it.
PASS internals.visualViewportScrollY() is 0
PASS internals.visualViewportScrollY() became 200
PASS successfullyParsed is true

TEST COMPLETE
